add: documentation. update: mintnft, routes
