+++ lib/flutterchain_lib/models/chains/near/mintbase_category_nft.dart
