+++ mintbase_example/lib/modules/pages/init/page/widgets/near_account_import_mnemonic.dart
@@ -4,48 +4,40 @@ import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:flutterchain/flutterchain_lib/models/chains/near/near_blockchain_data.dart';
+import 'package:flutterchain/flutterchain_lib/services/chains/near_blockchain_service.dart';
 import 'package:mintbase_example/consts/secure_storage_keys.dart';
 import 'package:mintbase_example/modules/models/authorized_model/authorized.dart';
 import 'package:mintbase_example/modules/models/auth_info_model/auth_info.dart';
 import 'package:mintbase_example/modules/controllers/auth_controller.dart';
-import 'package:mintbase_example/routes/routes.dart';
 import 'package:mintbase_example/widget/rounded_text_field.dart';
 
-class NearAccountImportActionDialogDefault extends StatefulWidget {
-  const NearAccountImportActionDialogDefault({
+class NearAccountImportMnemonic extends StatefulWidget {
+  final NearNetworkType networkType;
+
+  const NearAccountImportMnemonic({
     super.key,
     required this.networkType,
   });
 
-  final NearNetworkType networkType;
-
   @override
-  State<NearAccountImportActionDialogDefault> createState() =>
-      _NearAccountImportActionDialogDefaultState();
+  State<NearAccountImportMnemonic> createState() =>
+      _NearAccountImportMnemonicState();
 }
 
-class _NearAccountImportActionDialogDefaultState
-    extends State<NearAccountImportActionDialogDefault> {
+class _NearAccountImportMnemonicState extends State<NearAccountImportMnemonic> {
   bool loading = false;
 
-  final TextEditingController accountIdController = TextEditingController();
-  final TextEditingController privateKeyController = TextEditingController();
+  final TextEditingController mnemonicController = TextEditingController();
+  final TextEditingController acountIDController = TextEditingController();
 
-  void accountCredentialsValidation() {
-    if (accountIdController.text.isEmpty) {
-      throw Exception('Account ID is required');
+  Future<NearBlockChainData> getAccountInfo({required String mnemonic}) async {
+    if (mnemonicController.text == null) {
+      throw Exception("Input mnemonic");
     }
-    if (accountIdController.text.length < 6) {
-      throw Exception('Account ID isn\'t valid');
-    }
-
-    if (privateKeyController.text.isEmpty) {
-      throw Exception('Private key is required');
-    }
-
-    if (privateKeyController.text.startsWith('ed25519:')) {
-      throw Exception('Private key isn\'t valid');
-    }
+    final passphrase = "";
+    return await Modular.get<NearBlockChainService>()
+        .getBlockChainDataFromMnemonic(mnemonic, passphrase);
   }
 
   @override
@@ -58,28 +50,33 @@ class _NearAccountImportActionDialogDefaultState
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
+          SizedBox(height: 30.h),
           RoundedTextField(
-            textEditingController: accountIdController,
-            labelText: "Account ID",
+            textEditingController: acountIDController,
+            labelText: "AcountID",
           ),
-          SizedBox(height: 25.h),
+          SizedBox(height: 30.h),
+          SizedBox(height: 30.h),
           RoundedTextField(
-            textEditingController: privateKeyController,
-            labelText: "Private Key",
+            textEditingController: mnemonicController,
+            labelText: "Mnemonic",
           ),
           SizedBox(height: 30.h),
           if (!loading)
             FilledButton(
               onPressed: () async {
-                accountCredentialsValidation();
                 setState(() {
                   loading = true;
                 });
 
+                final accountInfo =
+                    await getAccountInfo(mnemonic: mnemonicController.text);
+
                 await Modular.get<AuthController>(key: "AuthController")
-                    .loginDefault(
-                        accountId: accountIdController.text,
-                        secretKey: privateKeyController.text,
+                    .loginWithMnemonic(
+                        accountId: acountIDController.text,
+                        publicKey: accountInfo.publicKey,
+                        secretKey: accountInfo.privateKey,
                         networkType: widget.networkType);
 
                 final FlutterSecureStorage secureStorage =
@@ -88,8 +85,8 @@ class _NearAccountImportActionDialogDefaultState
                   key: SecureStorageKeys.info,
                   value: jsonEncode(
                     Authorized(
-                      accountIdController.text,
-                      privateKeyController.text,
+                      acountIDController.text,
+                      accountInfo.privateKey,
                       widget.networkType,
                     ),
                   ),
@@ -100,7 +97,7 @@ class _NearAccountImportActionDialogDefaultState
                   loading = false;
                 });
 
-                Modular.to.navigate("${CoreRoutes.auth}");
+                Modular.to.pushNamed("//home/auth");
               },
               child: Text(
                 'Import Near account',
